requirements.txt
asgiref==3.5.2
Django==3.0
django-adminlte3==0.1.6
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-mptt==0.13.4
Pillow==9.2.0
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
